Add Radix UI checkbox component and render tasks
This commit adds the Radix UI checkbox component and renders a list of tasks using it. The @radix-ui/react-checkbox package was installed and the Checkbox component was created by wrapping the Radix UI primitive components. The MyTasks component was also added to render sample task data using the new Checkbox. This improves the UI by adding a nice checkbox component for selecting tasks.